[NO-JIRA] Clean up constants
